# openapi.yaml
openapi: 3.1.0
info:
    title: Online Booking App API
    version: 1.0.0
    description: RESTful API for an online booking app using Express.js and Prisma
servers:
    - url: https://api.example.com/v1
paths:
...
    /users:
        get:
            summary: Get all users
            responses:
                "200":
                    description: Successful retrieval of users
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/User"
        post:
            summary: Create a new user
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserCreate"
            responses:
                "201":
                    description: User created successfully
                "400":
                    description: Bad request
    /users/{userId}:
        parameters:
            - in: path
              name: userId
              required: true
              schema:
                  type: string
        get:
            summary: Get user by ID
            responses:
                "200":
                    description: Successful retrieval of user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                "404":
                    description: User not found
        put:
            summary: Update user by ID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserUpdate"
            responses:
                "200":
                    description: User updated successfully
                "404":
                    description: User not found
                "400":
                    description: Bad request
        delete:
            summary: Delete user by ID
            responses:
                "200":
                    description: User deleted successfully
                "404":
                    description: User not found
...
    /properties:
        get:
            summary: Get all properties
            responses:
                "200":
                    description: Successful retrieval of properties
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Property"
        post:
            summary: Create a new property
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PropertyCreate"
            responses:
                "201":
                    description: Property created successfully
                "400":
                    description: Bad request
    /properties/{propertyId}:
        parameters:
            - in: path
              name: propertyId
              required: true
              schema:
                  type: string
        get:
            summary: Get property by ID
            responses:
                "200":
                    description: Successful retrieval of property
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Property"
                "404":
                    description: Property not found
        put:
            summary: Update property by ID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PropertyUpdate"
            responses:
                "200":
                    description: Property updated successfully
                "404":
                    description: Property not found
                "400":
                    description: Bad request
        delete:
            summary: Delete property by ID
            responses:
                "200":
                    description: Property deleted successfully
                "404":
                    description: Property not found
...
    /reviews:
        get:
            summary: Get all reviews
            responses:
                "200":
                    description: Successful retrieval of reviews
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Review"
        post:
            summary: Create a new review
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ReviewCreate"
            responses:
                "201":
                    description: Review created successfully
                "400":
                    description: Bad request
    /reviews/{reviewId}:
        parameters:
            - in: path
              name: reviewId
              required: true
              schema:
                  type: string
        get:
            summary: Get review by ID
            responses:
                "200":
                    description: Successful retrieval of review
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Review"
                "404":
                    description: Review not found
        put:
            summary: Update review by ID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ReviewUpdate"
            responses:
                "200":
                    description: Review updated successfully
                "404":
                    description: Review not found
                "400":
                    description: Bad request
        delete:
            summary: Delete review by ID
            responses:
                "200":
                    description: Review deleted successfully
                "404":
                    description: Review not found
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "a1234567-89ab-cdef-0123-456789abcdef"
        username:
          type: string
          example: "jdoe"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "johndoe@example.com"
        phoneNumber:
          type: string
          example: "123-456-7890"
        profilePicture:
          type: string
          example: "https://example.com/images/johndoe.jpg"
...
    Property:
      type: object
      properties:
        id:
          type: string
          example: "g9012345-67ef-0123-4567-89abcdef0123"
        title:
          type: string
          example: "Cozy Mountain Retreat"
        description:
          type: string
          example: "Experience tranquility in our cozy cabin situated on a serene mountain peak."
        location:
          type: string
          example: "Rocky Mountains, Colorado"
        pricePerNight:
          type: number
          example: 120.5
        bedroomCount:
          type: integer
          example: 3
        bathRoomCount:
          type: integer
          example: 2
        maxGuestCount:
          type: integer
          example: 5
        hostId:
          type: string
          example: "f1234567-89ab-cdef-0123-456789abcdef"
        rating:
          type: number
          example: 5
...
    Review:
      type: object
      properties:
        id:
          type: string
          example: "g7890123-45cd-ef01-2345-6789abcdef01"
        userId:
          type: string
          example: "a1234567-89ab-cdef-0123-456789abcdef"
        propertyId:
          type: string
          example: "z0123456-78ab-cdef-9012-3456789abcdef"
        rating:
          type: number
          example: 5
        comment:
          type: string
          example: "The property was amazing, and the host was very accommodating!"
    UserCreate:
      type: object
      properties:
        username:
          type: string
          example: "jdoe"
        password:
          type: string
          example: "password123"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "johndoe@example.com"
        phoneNumber:
          type: string
          example: "123-456-7890"
        profilePicture:
          type: string
          example: "https://global-uploads.webflow.com/5eecfecbe625d195e35b89f2/624bfb093da7d92733c001c0_Ignacio%20Villafruela%20Rodr%C3%ADguez.jpg"
      required:
        - username
        - password
        - name
        - email
        - phoneNumber
    UserUpdate:
      type: object
      properties:
        username:
          type: string
          example: "jdoe_updated"
        name:
          type: string
          example: "John Doe Updated"
        email:
          type: string
          example: "updatedjohndoe@example.com"
        phoneNumber:
          type: string
          example: "987-654-3210"
        profilePicture:
          type: string
          example: "https://updated-example.com/images/johndoe.jpg"
...
    PropertyCreate:
      type: object
      properties:
        title:
          type: string
          example: "Cozy Mountain Retreat"
        description:
          type: string
          example: "Experience tranquility in our cozy cabin situated on a serene mountain peak."
        location:
          type: string
          example: "Rocky Mountains, Colorado"
        pricePerNight:
          type: number
          example: 120.5
        bedroomCount:
          type: integer
          example: 3
        bathRoomCount:
          type: integer
          example: 2
        maxGuestCount:
          type: integer
          example: 5
        hostId:
          type: string
          example: "f1234567-89ab-cdef-0123-456789abcdef"
        rating:
          type: number
          example: 5
    PropertyUpdate:
      type: object
      properties:
        title:
          type: string
          example: "Updated Cozy Mountain Retreat"
        description:
          type: string
          example: "Updated description for the cozy mountain retreat."
        location:
          type: string
          example: "Updated Location"
        pricePerNight:
          type: number
          example: 150.5
        bedroomCount:
          type: integer
          example: 4
        bathRoomCount:
          type: integer
          example: 3
        maxGuestCount:
          type: integer
          example: 6
        rating:
          type: number
          example: 4.5
...
    ReviewCreate:
      type: object
      properties:
        userId:
          type: string
          example: "a1234567-89ab-cdef-0123-456789abcdef"
        propertyId:
          type: string
          example: "z0123456-78ab-cdef-9012-3456789abcdef"
        rating:
          type: number
          example: 5
        comment:
          type: string
          example: "The property was amazing, and the host was very accommodating!"
    ReviewUpdate:
      type: object
      properties:
        rating:
          type: number
          example: 4
        comment:
          type: string
          example: "Updated review for the property."
...